move current matrix to math/static_matrix
next step, renaming it
