+++ libs/math/include/nil/crypto3/math/static_matrix/vector.hpp
