+++ libs/math/test/static_matrix.cpp
@@ -23,16 +23,15 @@
 // SOFTWARE.
 //---------------------------------------------------------------------------//
 
-#define BOOST_TEST_MODULE constexpr_vector_test
+#define BOOST_TEST_MODULE constexpr_matrix_test
 
 #include <boost/test/unit_test.hpp>
 #include <boost/test/data/test_case.hpp>
 #include <boost/test/data/monomorphic.hpp>
 
-#include <nil/crypto3/math/vector/vector.hpp>
-#include <nil/crypto3/math/vector/math.hpp>
-#include <nil/crypto3/math/vector/operators.hpp>
-#include <nil/crypto3/math/vector/utility.hpp>
+#include <nil/crypto3/math/static_matrix/matrix.hpp>
+#include <nil/crypto3/math/static_matrix/math.hpp>
+#include <nil/crypto3/math/static_matrix/operators.hpp>
 
 #include <nil/crypto3/algebra/fields/goldilocks.hpp>
 
@@ -41,6 +40,7 @@ using namespace nil::crypto3::math;
 
 using field = fields::goldilocks;
 using value = field::value_type;
+
 
 static_assert(make_vector(1, 2, 3) == vector {1, 2, 3}, "make_vector and uniform initialization deduction guide");
 
@@ -69,3 +69,42 @@ static_assert(generate<4>([](auto i) { return value(i * i); }) == vector<value, 
 static_assert(vector {1, 2, 3} == slice<3>(vector {1, 2, 3, 4}), "slice-no offset");
 
 static_assert(vector {2, 3, 4} == slice<3>(vector {1, 2, 3, 4}, 1), "slice with offset");
+
+
+// Uniform initialization
+constexpr matrix<value, 3, 3> m1 = {1, 2, 3, 4, 5, 6, 7, 8, 9};
+
+// Type deduction
+constexpr matrix m2 = {{{value(1), value(2)}}};
+
+constexpr matrix m22 = {{{value(1), value(3)}, {value(2), value(7)}}};
+
+static_assert(m1[0][2] == 3, "matrix[]");
+
+static_assert(m1.row(2) == vector {value(7), value(8), value(9)}, "matrix row");
+
+static_assert(m1.column(2) == vector {value(3), value(6), value(9)}, "matrix column");
+
+static_assert(fill<2, 2>(value(3)) == matrix {{{value(3), value(3)}, {value(3), value(3)}}}, "matrix fill");
+
+static_assert(matmul(m1, m1) == matrix<value, 3, 3> {{{30, 36, 42}, {66, 81, 96}, {102, 126, 150}}},
+              "real matrix multiply");
+
+static_assert(identity<value, 3> == matrix<value, 3, 3> {{{1, 0, 0}, {0, 1, 0}, {0, 0, 1}}}, "identity");
+
+static_assert(identity<value, 3> == inverse(identity<value, 3>), "inverse-identity");
+
+static_assert(inverse(m22) == matrix<value, 2, 2> {{{7, -3}, {-2, 1}}}, "inverse");
+
+static_assert(matmul(inverse(m22), matrix<value, 2, 1> {{{1}, {1}}}) == matrix<value, 2, 1> {{{4}, {-1}}},
+              "A^-1*b = x");
+
+static_assert(horzcat(identity<value, 2>, identity<value, 2>) == matrix<value, 2, 4> {{{1, 0, 1, 0}, {0, 1, 0, 1}}},
+              "horzcat");
+
+static_assert(submat<2, 2>(m1, 1, 1) == matrix<value, 2, 2> {{{5, 6}, {8, 9}}}, "submat");
+
+static_assert(rref(m1) == matrix<value, 3, 3> {{{1, 0, -1}, {0, 1, 2}, {0, 0, 0}}}, "rref");
+
+static_assert(rank(m1) == 2, "rank");
+
